requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.20
botocore==1.35.20
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.10
jmespath==1.0.1
multidict==6.1.0
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
RapidFuzz==3.10.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
yarl==1.11.1
